workers/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Vector embeddings
sentence-transformers==2.2.2
torch==2.1.0
openai==1.3.7
numpy==1.24.3
scikit-learn==1.3.2

# Patent processing
lxml==4.9.3
beautifulsoup4==4.12.2
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Search and indexing
opensearch-py==2.3.2
redis==5.0.1
nats-py==2.6.0
rank-bm25==0.2.2

# Cloud storage
boto3==1.34.0
minio==7.2.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
tenacity==8.2.3
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Document generation
python-docx==0.8.11
reportlab==4.0.4

# Graph analysis
networkx==3.2.1

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-asyncio==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0
prometheus-client==0.19.0
structlog==23.2.0

# Security
PyJWT==2.8.0
cryptography==41.0.7

# Error tracking
sentry-sdk==1.38.0
